Remove test log only if existing
As 'file.remove' breaks with an error if the file to be removed does not
exist, we now check for its existence first.

+++ test.R
@@ -6,7 +6,7 @@ source("util-plot.R")
 
 library(logging)
 logging::basicConfig(level = "DEBUG")
-file.remove("test.log")
+if (file.exists("test.log")) file.remove("test.log")
 logging::addHandler(logging::writeToFile, file = "test.log", level = "DEBUG")
 assign("last.warning", NULL, envir = baseenv())
 options(mc.cores = 6L)
